fix: Do not write to a tempfile
Writing to a tempfile and then moving it breaks using e.g. /dev/null
as output and probably other such things.

This replaces it with a destructor which will clean up the partially
written file unless explicitly marked as completed.  Reliability-wise
this is as good as they tempfile solution as tempfile did also rely on
drop guards.