Only allow one request per stream
The only way to be able to emit the event that a transfer is completed
while knowing the client received all data is by closing the
SendStream.  This requires the protocol to be changed since now only a
single request can be sent per stream.

Additionally this exposed that a test was stopping the provider too
soon.  The sequencing of quinn here is perhaps a bit peculiar, but
that's how it is.