Call finish on the SendStream, not stop on RecvStream
It turns out that calling SendStream::finish is the graceful way of
stopping sending data on a stream.  It simply sets a bit on the last
STREAM frame sent and that's it.

Calling RecvStream::stop on the other hand is the only way for the
receiving endpoint to reset a stream.  It results in an extra
STOP_STREAM frame to the sender which in turn will reset the stream
and send a RESET_STREAM frame to the receiver.  Not great.